+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version header format: `## [x.x.x] - yyyy-mm-dd`
 
 - Fixed code scan stderr/stdout collision by moving the stdout dump to the end of the run function
 - Fixed image scan stderr/stdout collision by moving the stdout dump to the end of the run function
+- Fixed image build disable check
 
 ## [v0.0.0-rc.12]